+++ src/screens/tagManagement/components/TagList.tsx
@@ -1,0 +1,73 @@
+import React, { useState } from 'react'
+
+import { useMutation } from 'react-query'
+
+import { deleteTag } from '@/apis/tag'
+import TwoButtonDialog from '@/components/dialogs/twoButtonDialog/TwoButtonDialog'
+import i18n from '@/locales'
+
+import {
+  ButtonText,
+  DeleteButton,
+  GrayDivider,
+  TagListContainer,
+  Text,
+} from '../TagManagement.style'
+
+interface TagListProps {
+  id: number
+  name: string
+  editMode: boolean
+}
+
+/**
+ * 태그 관리 리스트
+ */
+export const TagList = ({ id, name, editMode }: TagListProps) => {
+  const [isDeleteDialogVisible, setIsDeleteDialogVisible] = useState(false)
+
+  const { mutate: deleteTagMutate } = useMutation(deleteTag)
+
+  /**
+   * 선택한 tag를 삭제합니다.
+   */
+  const handleDelete = (tagId: number) => {
+    deleteTagMutate(tagId)
+  }
+
+  return (
+    <>
+      {editMode ? (
+        <>
+          <TagListContainer key={id}>
+            <Text>{name}</Text>
+            <DeleteButton onPress={() => setIsDeleteDialogVisible(true)}>
+              <ButtonText>{i18n.t('delete')}</ButtonText>
+            </DeleteButton>
+          </TagListContainer>
+          <GrayDivider />
+          <TwoButtonDialog
+            isVisible={isDeleteDialogVisible}
+            title="doYouWantDeleteThisTag"
+            description="deleteTagsCannotBeRestored"
+            completeText="delete"
+            onCancel={() => {
+              setIsDeleteDialogVisible(false)
+            }}
+            onComplete={() => {
+              setIsDeleteDialogVisible(false)
+              handleDelete(id)
+            }}
+          />
+        </>
+      ) : (
+        <>
+          <TagListContainer key={id}>
+            <Text>{name}</Text>
+          </TagListContainer>
+          <GrayDivider />
+        </>
+      )}
+    </>
+  )
+}
